Bug #25153670: INCORRECT BEHAVIOR WITH MULTI-TABLE UPDATE                USING TEMPORARY TABLE
Issue:
------
This problem occurs under the following conditions:
1) When an UPDATE requires a temporary table.
2) This temporary table is then created on INNODB.
3) The size of the target table's primary key is greater
   than 1024.

Root cause:
-----------
The temp table is used to hold the rows that qualify and
need to be updated. This is done by storing the target
table's primary key (PK) in the temp table to identify the
row. All the rows are identified and stored in temp table.
Then PK values in the temp table are used to identify the
rows in t1 (target table) and are then updated.

While creating a temp table,
Query_result_update::optimize() creates a single
field (Field_string) to squeeze all the columns of the
target table's PK. This works fine if the size of
PK < 1024.

INNODB's fixed length column size is restricted to 1024
(See "struct dict_field_t"). Exceeding this can result in
incorrect behavior.

Solution:
---------
Since we have been unable to identify a good fix for
this issue, an error will be returned for this DML
statement.

A different bug will be opened to track and fix this error.


Changes specific to trunk:
--------------------------
sql_update.cc:
While creating temp table, the big_tables option is unset.
This code has been removed. This means that optimizer never
chooses INNODB for temp table for UPDATE statements. The
only way INNODB temp tables come into the picture is when
there is an overflow from the temp table on HEAP.

with_non_recursive.result:
After the above change, some UPDATE statements with
big_tables option will use INNODB temp tables.