Bug#26730020 REEXECUTE PREPARED STATEMENT CRASH WITH WINDOW FUNCTIONS
When setting up windows' "PARTITION BY/ORDER BY"s caches
(cf. setup_ordering_cached_items) we need the "PARTITION BY/ORDER BY"s
expressions to be resolved first (cf. resolve_window_ordering).

On a first prepare this happens in two different loops over the
windows, i.e.  all resolving is done before we proceed to the cache
setup.

On an explicit (re)execute, another code path is taken, which do these
two steps inside one loop. Normally this is fine, but in the repro it
is not, consider our window definitions:

   WINDOW w2 AS (w1), w1 AS (ORDER BY a, a)

We process the windows in textual order. Note that the ordering for w2
is inherited from w1. That is, the cache setup for w2 depends on the
order by for w1 having been resolved first (it is not done as part of
processing w1 since it doesn't have an ORDER BY clause, and we only
want to do the resolution once). The result is that we try to set up
the cache for w2 *before* resolution of its effective (i.e. inherited)
ORDER BY clause. This causes the issue.

Fix: do all resolution first also on a (re)execute.

Repro test added.