Bug#23021693 gcol:assertion '!table->in_use->is_error()' failed
The problem here is that an error leaks from test_quick_select()
into the executor without being correctly processed.
When update_generated_read_fields() is called, it expects no current
error state and asserts.

It is a major problem that subsystems like the range optimizer has no
proper way of propagating an error state, however this is a minimum
fix that checks for error immediately after the call and terminates
execution when an error state is encountered.