Bug#26802696 DENSE_RANK WRONG WITH BUFFERED PROCESSING (follow-up)
It turns out "Item_percent_rank" has the same logic error as
"Item_rank", but we can fix this by taking advantage of the fact this
is a two-pass window function, so we can inquire
"Window::last_rowno_in_peerset()" instead of comparing with previous
row (as is done for RANK and PERCENT_RANK) in the window function's
own logic, since they are not two-pass. Aside: this was not done
originally because the "last_rowno_in_peerset" state hadn't been
introduced yet.

With this patch, we now set "needs_peerset" to true in
"Item_percent_rank::check_wf_semantics" and use
"last_rowno_in_peerset()" to calculate the rank there by eliminating
m_previous member.

Change-Id: Ifa1d2d827fcbd4164234740c91f5058d8fcdfee0