Bug#26659699 Assertion !(thd->lex->context_analysis_only & 2) failed
The assert is hit in a CREATE VIEW statement on partitioned tables,
when the server is trying to prune partitions of the underlying tables.
A const subquery is attempted executed, but hits an assert in
Item_subselect::exec(), since no subqueries should be evaluated
when just analyzing a view creation.

Partition pruning is called from SELECT_LEX::apply_local_transforms(),
however, it is not reasonable to apply any transformations when
analyzing a view creation, so the simple solution to this problem
is to bail out of apply_local_transforms() immediately when analyzing
a view creation.