BUG#26629969: FAILURE IN I_INNODB.TABLE_COMPRESS_3 AFTER WL#9538
Description:
------------
The test i_innodb.table_compress_3 is failing with '--mem' option
after the push of wl#9538. When '--mem' option is enabled, MTR sets
the value of innodb_use_native_aio mysqld variable to 0 and this
change is causing the result diff in the test run.

Fix:
----
Remove the code from MTR which checks for '--mem' option and sets the
value of 'innodb_use_native_aio' to 0 since its not needed anymore.

Reviewed-by: Satya Bodapati <satya.bodapati@oracle.com>
RB: 17314