Bug#23529242 CONCAT: ASSERTION FAILED: MAYBE_NULL
Analysis:

In the repro:

   SET SQL_MODE= ''; // required to see issue, see [1]
   SELECT CONCAT(UNIX_TIMESTAMP(COUNT(1)), '|');

we hit an assert that Item_func_concat::maybe_null is true when we see
a NULL being returned from UNIX_TIMESTAMP.

UNIX_TIMESTAMP return a NULL due to the illegal
"datetime" value:

    mysql> select unix_timestamp(count(1));
    +--------------------------+
    | unix_timestamp(count(1)) |
    +--------------------------+
    |                     NULL |
    +--------------------------+
    1 row in set, 2 warnings (0.00 sec)

    Warning (Code 1292): Incorrect datetime value: '1'

TIMESTAMP(1) and UNIX_TIMESTAMP(1) do behave differently:

    mysql> select timestamp(1);
    +--------------+
    | timestamp(1) |
    +--------------+
    | NULL         |
    +--------------+
    1 row in set, 1 warning (0.00 sec)

    mysql> select unix_timestamp(1);
    +-------------------+
    | unix_timestamp(1) |
    +-------------------+
    |                 0 |
    +-------------------+
    1 row in set, 1 warning (0.00 sec)

and the documentation for UNIX_TIMESTAMP states unequivocally:

  "If you pass an out-of-range date to UNIX_TIMESTAMP(), it returns 0."

But on the other hand when we have the COUNT(1) aggregate as an
argument we get a NULL, which is a bug.

Next, CONCAT has deduced it is not nullable based on the information
about its arguments: UNIX_TIMESTAMP and the literal which is trivially
not NULL. So, when that function does return a NULL,
Item_func_concat::val_str hits the ASSERT.

CONCAT has the wrong idea: it is indeed nullable because
UNIX_TIMESTAMP is so, and the latter isn't marked as such. If it were,
CONCAT would get that property too by the propagation in
Item_func::fix_func_arg, but see [1].

But back to COUNT(1): why doesn't the wrong argument COUNT(1), which
evaluates to 1, give the same result as plain 1 as an argument.  The
crucial difference is an intervening Item_ref in the item tree for the
former which is sets null_value if an error is returned from the
called get_date (called from Item_unix_timestamp::val_timeval ->
Item::get_time_val -> Item_ref::get_date:

   return (null_value= (*ref)->get_date_result(ltime,fuzzydate));

so in turn, UNIX_TIMESTAMP get its null_value set, too.  We see,
though, that Item_unix_timestamp::val_timeval doesn't expect any
underlying Item_ref to do just that (from its Doxygen):

/**
   If argument is NULL, sets null_value. Otherwise:
   if invalid DATETIME value, or a valid DATETIME value but which is out of
   the supported Unix timestamp range, sets 'tm' to 0.
*/

Fix:

Let UNIX_TIMESTAMP(1) continue to return 0 when it sees an invalid
argument, in contrast to TIMESTAMP().
We can achieve this by letting Item_ref::get_date just return the error
code of the underlying get_date_result as before, but explicitly propagate
the null value (instead on setting it only based on the error code), thus:

  bool result= (*ref)->get_date_result(ltime,fuzzydate);
  null_value= (*ref)->null_value;
  return result;

Note that Item_direct_ref::get_date already does the null propagation in this way.

With the change, time_func with the new repro passes. I will run a
larger set of tests to see if it has any repercussions.

[1]

If we use strict mode, CONCAT is marked as nullable disregarding its arguments
in Item_str_func::fix_fields, cf this line, last condition:
  :
  /*
    In Item_str_func::check_well_formed_result() we may set null_value
    flag on the same condition as in test() below.
  */
  maybe_null= (maybe_null || thd->is_strict_mode());
  :

This is why the setting of SQL_MODE is required in the repro.

Repro added in func_time.test.

Change-Id: Idd552856465ababd95a6c40794d67a57efb4b7ed