Bug#26780307: INCONSISTENT JSON ESCAPING WHEN USING JSON TYPE
The JSON specification says that all control characters from U+0000 to
U+001F need to be escaped, but when MySQL shows JSON data as text, it
does not escape the character U+001F.

The fix is to escape all characters less than or equal to 0x1f.
Previously it only escaped those less than 0x1f.

This is a partial backport of bug#26388690.

Change-Id: Ia9b4e6653057ac817ab0ca2e0f9c1895ad0f637e