Bug#25658967 REVOKE ALL IS NOT REVOKING ALL PRIVILEGE WHEN CURRENT_USER() IS USED
The REVOKE statement failed to revoke dynamic privileges if the
target user was specified as CURRENT_USER().
CURRENT_USER() is not evaluated by the parser. Instead evaluation
must happen at runtime for each and every statement which use this
function which wasn't the case for the code related to revolval of
all dynamic privileges.