Bug #26781567: REMOVE C LEGACY [noclose]
Rename most st_* structs with a typedef to just being that typedef;
there's no point in a struct having two names. There are still a few
remaining, which will be dealt with in a follow-up patch.

structs that are expected to be used from C code (e.g. mysql.h) have
been given a typedef to themselves, so that C code can continue to
write MYSQL *mysql instead of struct MYSQL *mysql.

Change-Id: Ic4bdd6cd9e2d9d1223c331ddd1ad89423742fba2