Bug#26758939 : FAIL TO LOCK THREAD TO CPU
When an ndbd is configured to lock its thread to its CPUs, it crashes
during start phase with a "Failed to lock tid" error message.

This bug is a side effect of fix for Bug#26378589. That patch disables
the cpu locking feature if the required glibc version is not available.
The bug in the glibc will be hit only during a Ndb_UnlockCPU() call.
The datanodes never use this call. This unlock call can be only
accessed through the API calls. So, this patch enables the cpu locking
feature for the datanodes and disable them only for the API calls,
based on the glibc version used. This patch also adds restrictions to
prevent any future uses of sched_setaffinity() call with an all ones
CPU mask in an unsupported glibc environment. The relevant ClusterJ
message is also updated to provide more information to the user.

(cherry picked from commit 9dd6bdd9915ea2b94d15c34e8bc36ef7078cd8c5)