WL#11139: Remove group_replication_allow_local_disjoint_gtids_join option
The option group_replication_allow_local_disjoint_gtids_join[1] was
introduced with the purpose[2]:
  This variable is an override mechanism for a new consistency feature
  in the plugin. When you try to add to the group a server that has
  more data than the existing members, by default the joining member
  will be rejected.
  The purpose is to avoid possible recovery and run time errors in
  the plugin leading to data divergence and other issues. If you are
  sure of what you are doing, and that the extra data is safe, then
  you can use this option.

The scenarios on which this can help users are:

 S1) Solve a broken majority asap
     * Group had 3 members in single-primary mode
     * 2 members crashed and end up with different data
       (GTID_EXECUTED) when compared with the group.
     * User wants to add one of the members (S2) to solve the
       majority loss or increase group members number to handle
       possible future member failures.
     * The member that is still on the group is the one that is the
       primary or will be elected as primary when majority is
       re-established, since is the only one ONLINE.
     * The group will be stable if a future failover does not switch
       to the S2 member.

The runtime errors that may happen are that the member which did
force the join with disjoint GTIDS, if it does a write, the changed
rows will never be updated by other members. Certification will
fail.
The recovery errors that may happen are that if this member is
chosen as donor, it will distribute the disjoint GTIDS among the new
members and increase the likelihood of the runtime errors to happen.

It must never be used unless the DBA knows exactly what she/he is
doing or does need to solve a broken majority, on which crashed
members suffered corruption and she/he will not perform writes on
it. Which can already be done by resetting master and setting
GTID_PURGED to match the GTID executed of the group if they really
want to add the damaged server to the group.

People use this option to force member join, and assume that since
member did successfully join everything is OK. Which is not the
case, data is inconsistent.

The correct procedure to fix a majority loss is described at
https://dev.mysql.com/doc/refman/8.0/en/group-replication-network-partitioning.html

Since the trade-off between this option benefit vs danger is
negative, we are removing this option.