Bug#26333981 BI IS WRITING TABLE_MAP EVEN WHEN USING SQL_LOG_BIN = 0
Added support for maintaining a accumulative any_value bitmap
for GCI event operations that represents bits set consistently
for all rows of a specific table in an epoch. A new method variant
Ndb::getNextEventOpInEpoch3 is added that allows for retrieving the
cumulative any_value bitmap.
Added a check if all operations (rows) for a specific tables are all
marked as nologging to exclude adding this table from the Table_map
of the binlog injector. Added a new test case that verifies this case
by checking just the Table_map (independent of how epochs are generated).