BUG#26802395: ASSERT `SERVER_ENGINE_INITIALIZED()' AT DELAYED_PLUGIN_INITIALIZATION.CC:167
During server initialization, with --initialize or
--initialize-insecure, replication applier infrastructure is not
initialized, as such Group Replication cannot be started on server
boot.

The above condition was not being properly considered, which led to
a assertion failed.
To avoid that issue, now the condition is properly considered and
Group Replication is not started, which is intended since the server
is being initialized.