Bug #26627181: WRONG RESULT WITH LEFT JOIN + DERIVED TABLES
Issue:
------
This bug can occur when:
- there is an outer join.
- an Item_direct_view_ref object references a const value
  from an inner table.

The problem occurs on latest 5.7.20 but not on 8.0. The
reason is that the fix for Bug#26432173 was only a
partial fix in 5.7 (i.e. fixed only for temp table cases)
but a more general one in 8.0.

Solution:
---------
A new Item_direct_view_ref::used_tables() function is
created to take into account whether an Item in a
derived table / view is part of an inner table of an outer
join.

The new tests have been added to 8.0 and trunk.