Bug #26731025	INNODB PARTITION TABLE WILL LOCK INTO THE NEAR RECORD 		AS A CONDITION IN THE USE
PROBLEM
=======
During a multiple client scenario:
Multiple updates from diffrent sessions end up in lock wait timeout
issue.
The lock wait timeout is occuring because when session 1 tries to update
a record and takes a record lock on it and other session tries doing
another update on another record, it starts to wait for lock to be
released by session 1 when it tries reading the next record in range.
Session 2 while fetching the next record calls on index_next that sets
the match mode to be 0 instead of ROW_SEL_EXACT or 1 which would
correctly set the lock type to be GAP LOCK and avoid the uneccessary
wait it would have to do otherwise in the case when it can't have
explict lock on record.

FIX
====
Its a regression from 5.7 onwards, in 5.6 the issue doesn't exist.
Have made the call to index_next_same instead of index_next only in the
case when scan type is PARTITION READ RANGE and the bool value of
is_next_same is set to true, so that match mode is correctly set and
LOCK_GAP gets set avoiding the uneccessary wait.

Reviewed by: Debarun Banerjee <debarun.banerjee@oracle.com>
RB: 17380