Bug #26730000 	SERVER_ID LOOKS SET TO VALID VALUE STILL START SLAVE FAILS WITH ER_BAD_SLAVE
Now, default server_id is set to 1 and user gets a warning
if it is not set during server initialization. But, there
is no WARNING and starting slave fails when server-id is
not supplied by users during server starts.

To fix the problem, remove the check on server_id_supplied
on starting slave, and report a warning if the server-id
is not supplied by users during server starts.