Bug #26826272: REMOVE GCC 8 WARNINGS [noclose]
Fix an antipattern in my_load_defaults where the MEM_ROOT was stored
in a memory just before the (new) argv array; replace it with an explicit
MEM_ROOT, which has a much cleaner destructor pattern.

The main motivation for this is to get rid of the GCC 8 warning about
memcpy-ing the MEM_ROOT.

Re-push: Fixed a crash in the NDB management daemon.

Change-Id: Ib6d09dac915020ef6413464fc9adea437c22185e