BUG#26378948: ASSERTION IN GROUP_REPLICATION.GR_FILTER_CRASH
The "--shutdown_server 0" at the test case is sometimes (depending
on the host overload) making the debug framework to return from
"debug_sync_set_action" function with an error (because no other thread
signaled the "signal.continue").

The debug sync was used to block server 1 to hold server 2 on recovery,
now we stop group_replication_applier on server 1 to cause same behavior.