BUG#26266789: ON SERVER 8.0 CREATE OR REPLACE VIEW               FAILS FOR COLUMN SIZE > 64
Analysis
========
Creating a view with explicit view column name greater
than 64 characters reports an error on release build
and triggers an assert on debug build.

In 5.7, a view can be created with explicit column
name greater than 64 chars. In 8.0, the column name
in DD::COLUMNS is restricted to 64 characters(maximum
column name length). Hence while attempting to store
a view column name which exceeds 64 chars an assert
is triggered.

Fix:
===
If the view column name exceeds the maximum column
name length(64 chars), an appropriate error is reported.
This makes it align with the restriction of 64 characters
for aliases of column names in CREATE VIEW statements.