Bug#26773152 - ASSERTION FAILURE: DICT0DD.CC:1154:FORM->S->REAL_ROW_TYPE == ROW_TYPE_COMPRESSED
This is a bogus assertion. It was expected that if key_block_size exists
but the real type read from se_private_data is not COMPRESSED,
then it must be a partitioned table. However this is not true.
It's possible that a table with key_block_size = 16K and
row_format = COMPRESSED can be reorganized into a tablespace of
format DYNAMIC etc. by an ALTER TALBE ... COPY. After this ALTER TABLE,
the row format of this table would become the same with the tablespace.
So it's not true the real_row_type must be ROW_TYPE_COMPRESSED.

Solution is to remove this bogus assertion.

RB: 17414
Reviewed-by: Allen Lai <allen.lai@oracle.com>