Bug#26126789: lock0lock.cc: ((m_mode - (0xful & m_mode)) - (0xf0ul & m_mode) - (257 & m_mode))
Problem: Failed creation of a temporary table using REPLACE could in
certain circustance lead to an assert in a later statement.

Root cause was that the trx_t::duplicates member is only properly reset
when calling handler::ha_reset(). This function is not guaranteed to be
called at the end of each statement, but usually it is.

Fix: Make Query_result_create::drop_open_table() call handler::ha_reset()
explicitly for temporary tables. For normal tables close_thread_table()
is called which already calls ha_reset().

This fixes the observed problem, but note that there are other
situations where a handler is closed without calling ha_reset().