Bug#26678883 - ASSERTION `HAS_ERROR == THD->GET_STMT_DA()->IS_ERROR()
The assertion means that InnoDB ever pushed an error when the thread was
interrupted however ha_innobase::write_row didn't return an error to server.
It's the asynchronous rollback, who required the MDL on related tables,
received the interrupted during requesting MDL and pushed the error.
It's difficult to check every operation after it killes other transactions,
and the key is that for the asynchronous rollback, MDL is totally unnecessary.

So the fix is in dd_mdl_for_undo(), the asynchronous rollback should be
screened out and no mdl would be required. Also in this patch, the problem
that only in-memory tables would be checked during opening tables for undo
has been fixed. To open the table, it's wrong to assume the table is always
in memory, so global DD would be accessed too.

RB: 17422
Reviewed-by: Jimmy Yang <jimmy.yang@oracle.com>