Bug#25820175 - i_s query pushes down function as icp and opens dictionary table
INFORMATION_SCHEMA views TABLES and STATISTICS uses internal
functions which opens dictionary tables. When ICP is enabled, if
these functions are used in the where condition then they are
pushed down to the storage engine for the evaluation. But these
internal functions should not be pushed to the storage engine.

To fix this issue, marked internal function opening data dictionary
table as DD_INTERNAL_FUNC. If ICP is enabled then while evaluating
condition to push down, internal functions of type DD_INTERNAL_FUNC
are not pushed down.

Patch for this issue tested using the RQG.

Change-Id: Ib3501fd923b149581c22261122f19b0bf6695812