wl#11164 Extend TreeNode class with member representing 'order of execution'.
In this patch we introduce member variables to TreeNode class
which represent the 'order of execution' among the TreeNodes.

The 'order of execution' is temporarily set up to
be the same as the child/parent relationship as currently defined
by the API in the SPJ request, so no functional changes yet.

Future intention is to use this in SPJs own query planner
to make local choices about execution order.