Bug#26324852: THERE IS NO LOCAL IP ADDRESS MATCHING THE ONE CONFIGURED FOR THE LOCAL NODE
Problem:
========
We bind many virtual IPs in one machine, it seems that group
replication can not recognize all of the virtual IPs that we
bind.

Execution of "start group_replication" will fail and error
log shows :

'[GCS] There is no local IP address matching the one
configured for the local node (10.202.44.222:24901).

Analysis:
=========
No matter how many virtual IPs that users configure, MySQL
group replication will be able to fetch only the first 12
IPs among them. The reason for that is, within the code an
initial buffer of size 512 bytes is allocated to hold list
of network interfaces. Each interface is represented by
'struct ifreq'. The size of 'struct ifreq'= 40 bytes.

Hence the buffer that was provided by the group replication
layer can hold a maximum of 12 interfaces.

i.e (40 *12)=480 < 512.

Documented behavior of 'ioctnl':
================================
'ioctnl' function is used to retrieve the list of interface
addresses. The user will be passing a buffer and the kernel
fills the ifreqs with all current L3 interface addresses
that are running. Along with this the kernel also returns
actual length in ifc_len. If ifc_len is equal to the
original length the buffer probably has overflowed and user
should retry with a bigger buffer to get all addresses.

In order to do an appropriate overflow check, the ifc_len
should be a multiple of struct ifreq size. Since the
initialized value '512' is not a multiple the code will
return earlier after reading the first 12 IP addresses.

Fix:
===
Changed the initial buffer size to be a multiple of struct
ifreq size.