wl#11164 addendum patch fixing a regression in ::storeRow()
A couple af the testSpj ATR testcases failed on the WL-branch
due to a regression introduced when extending the BUFFER functionality
on one of the previous patches.

As part of ::StoreRow(), the buffered row may either be registered
in a list or a (collationId-) map of rows. This process may
involve allocation of the map structure itself, which has the
potential for the entire memory area we does such allocation from
to be reallocated and moved.

Thus, the 'dstptr' which storeRow() write the buffered row into
effectively became invalid.

This patch avoid this problem by ensuring that buffering
of the row is completed before the add_to_map() / _list()
method is called. The 'dstptr' is not referred after this,
so we will not be affected by a realloc.