BUG#26562371 INNODB: ASSERTION FAILURE: LOCK0LOCK.CC:NNN:!TRX || !LOCK_REC_OTHER_TRX_HOLDS_EX
Issue:
======

After [ wl#10793: Use VATS for scheduling lock release under high load ], we
ignore transaction locks that are marked or chosen as deadlock victim.

This needs to be handled in debug functions where we assert for conflicting lock.
We handle it already in lock_rec_other_has_expl_req() and a similar check is
needed in lock_rec_other_trx_holds_expl().

Fix:
====

Fix for [Bug #24344131] - rb#16742, handles similar debug assert issue when a
transaction is perceived committed after it is removed from rw list.

So the fix for this bug involves putting together both the conditions and checking
in all required places to avoid assert for conflicting locks:

1)    Transaction is being committed [not in rw list]
2)    Transaction is being rolled back

RB: 17101
Reviewed-by: Debarun Banerjee <debarun.banerjee@oracle.com>