Bug#26178130: MTR REPORTS "USE OF UNINITIALIZED VALUE IN CONCATENATION"
Description
-----------
MTR sometimes reports following error on its console:
Use of uninitialized value in concatenation (.) or string at
<path>/mysql-test-run.pl line <line>.

Fix
---
The warning was being reported due to the property 'shortname'
being undefined in the test information hash for few reporting
tests like the one used for the --report-features option. The
property has been initialized with its name in order to resolve
the issue.

Reviewed-by: Pavan Naik <pavan.naik@oracle.com>
RB: 17320