BUG#26280724 ASSERT FAILURE WHEN REPLICATING TABLE WITH VIRTUAL JSON COLUMN
Problem and analysis:
When row based replication is enabled and slave-rows-search-algorithms
is set to 'INDEX_SCAN,HASH_SCAN', updating a table containing a virtual
generated field that relies on its string representation to generate
the hash for the table scan would result in an error when generating
the said string.

Fix:
Since virtual generated fields cannot be used to identify a row, do
not create hashes for virtual generated fields. Also, exclude these
fields when comparing records field by field.