Bug#21373532: RANGE TESTS FAILS DUE TO DIFFERENCE IN EXPLAIN OUTPUT
One of the test case change has already been pushed. This patch fixes
index_merge_delete.test

The variation in the plan is expected on different platforms as stated in
the test file. Added what is done for other tests just below.

Regarding the new change ( from "Delete all rows" to NULL
in the plan) for some of the deletes is due to changes in 8.0. The
condition (a subquery) is not a const_item() anymore which makes optimizer
choose a different plan in 8.0 (in 5.7, const_item() for the subquery
returns true).