Bug#26679495: SIG 11 IN SUBSELECT_HASH_SJ_ENGINE::CLEANUP|SQL/ITEM_SUBSELECT.CC
If a query uses hash semi-join to execute an IN subquery, and an error
is raised in subselect_hash_sj_engine::setup() (for example because
the disk is full), the server may crash while cleaning up after the
query, because some of subselect_hash_sj_engine's members are not
initialized.

The fix is to make the cleanup code check if the members are
initialized before it tries to clean them up.