BUG#26878022 MYSQLBINLOG: ASSERTION `(OLD_MH->M_KEY == KEY) || (OLD_MH->M_KEY == 0)' FAILED
Problem and Analysis
--------------------

The mysqlbinlog client program is trying to reallocate a buffer passing
a pointer to an advanced position of the buffer instead of the base
address of the buffer.

This is making P_S and realloc function to not detect correct PSI memory
key for the buffer.

Fix