Bug#26679983: SIG 11 IN MAKE_JOIN_READINFO|SQL/SQL_SELECT.CC
If an error happens when setting up the temporary table for duplicate
weedout in a semi-join, for example because the disk is full, the
server may terminate instead of gracefully giving an error message
describing the problem.

The fix is to check the return value from
create_duplicate_weedout_tmp_table() and abort the query early if
creating the temporary table failed.