Bug#26428017 VARIABLE SCOPE OF AUTOCOMMIT NOT CHANGING TO COMMAND_LINE/EXPLICIT/EXTRA
Issue:

Before this fix, table performance_schema.variables_info
was not reporting proper data for the autocommit system variable.

In particular, the VARIABLE_SOURCE column was always COMPILED,
even when the autocommit system variables set in command lines argument
or in configuration files.

Root cause:

The root cause is specific to the --autocommit variable itself

- Sys_autocommit is a Sys_var_bit
- my_handle_options() does not support *bit* values, only booleans,
  so that my_handle_options() can not set the autocommit bit in
  THD::variables.option_bit directly

To work around that, a *second* option, also named autocommit,
was added in my_long_options[]

As a result, there are *two* options named "autocommit",
which confuses the code when keeping track of the option source.

The --autocommit option source is not preserved (because it is not a system
variable), while the @@global.autocommit system variable source is never
updated (it is shadowed by --autocommit).

Fix:

Allocate a source structure for the --autocommit option,
so that the source of --autocommit is preserved when parsing arguments.

When the data is copied from --autocommit to @@global.autocommit,
also copy the metadata (the option source)