Bug #26435766: PASSWORD EXPIRE FLAG IS NOT WORKING PROPERLY FOR   EXTERNAL AUTHENTICATION PLUGINS
Two things:
1. Added a check if non-default password expiration options (e.g.
PASSWORD EXPIRE or PASSWORD EXPIRE INTERVAL 1 DAY) are passed to
CREATE/ALTER USER that will use an authentication method that does
not support password expiration. Note that PASSWORD EXPIRE DEFAULT
and PASSWORD EXPIRE NEVER are still supported.
2. Now the password expiration flag is reset for all authetnication
methods not only for the ones that support password expiration.
This is needed since there's no need to keep the flag on for methods
that do not support it.