BUG#26626277: BUG IN "INSERT... ON DUPLICATE KEY UPDATE" QUERY
Analysis:
=========
The AFTER UPDATE trigger is not invoked for 'INSERT..ON DUPLICATE
KEY UPDATE', when the value to be updated and the new value
provided are the same.

The AFTER UPDATE TRIGGER was triggered only if the new and old
record values were different for 'INSERT..ON DUPLICATE KEY UPDATE'
handling. Since the values specified in the UPDATE clause
and the old value were the same, the trigger was never invoked.

Fix:
====
The AFTER UPDATE trigger is invoked even if the values specified
in the UPDATE clause and the old value are the same.