Bug#26813454 UNWARRANTED ERROR INDICATING WINDOW NAME NOT DEFINED
The SELECT's window list gets corrupted, so the name resolution can't find the
window definition.

This happens because of a bug in the List class: for windows we use a combination
of List::prepand and List::push_back.

List::prepand is buggy and doesn't maintain the List::last invariant
is the list is empty. ::prepand is used for the set of explicit window
definitions.

When we later attempt to push_back the window definition in the ORDER BY,
the list gets corrupted.

This patch fixes this, corrects the spelling to ::prepend, adds a test
case to window_functions.test and a new unit test to sql_list-t.

Change-Id: Icf05d976b5d9bf8c6c0f6ec68f9535b275f72de2