BUG#26133488 SLAVE IS DIVERGED FROM MASTER BUT CROSS-VERSION RPL CONTINUES WITHOUT ERROR
Problem:
An 8.0+ slave supporting atomic DDL's may diverge from an older 5.7-
master without any warning: if a part of the failed DDL can not be
applied, then it is completely rollback by the slave, while the master
may have partially committed it, which may result in their states
having silently diverged.

Fix:
When applying a transaction without a valid xid defined by the master
(a 5.7- server that does not support atomic DDL's), if it is an atomic
DDL and its commit was not successful both on master and slave, an
error will be logged.
Also increased MAX_SLAVE_ERROR because the current maximum error
number is larger than the previously set 10000.