Bug#26802696 DENSE_RANK WRONG WITH BUFFERED PROCESSING
For the first row in a partition when we have buffering, we
erroneously get values that are one too high. RANK has the same issue,
but there is accidentally works correctly anyway.

Analysis:

The reason is that the initialization of the cache comparator of the
ORDER BY happens too early. In streaming mode, we call clear (which
resets the comparator) in Item_rank::val_int - which works fine because
the values of the ORDER BY expressions is already available in the
outbuffer (copy_fields has been done already).

In buffered mode, however, the call to clear happens from
process_buffered_windowing_record *before* copy_fields has been called
as part of a general reset of window functions' states (when we start
a partition).

Solution:

Split the state reset and the priming of the comparator, so
that the latter always happens in Item_rank::val_int.

Repro added.

Change-Id: I8157ed96f2aca7a843ebf0a80364c75d52adef48