Bug#26898576: --SYMBOLIC-LINKS/-S' IS DEPRECATED MESSAGE SHOULD BE MORE HELPFUL
Problem: Since the symbolic links option is deprecated, a warning was
printed whenever it was used. But since its use has previously been
discouraged, many users disable it explicitly, and they are given the
same warning.

Fix: Adjust the warning to inform users that the option is disabled by
default when it is being disabled explicitly. Issue a normal
deprecation warning when the option is being enabled.

Also move the test for !opt_help before the call to my_setwd() for
datadir. This prevents an warning about missing datadir from being
printed when executing --help without a valid datadir.