Bug #26251621: INCORRECT BEHAVIOR WITH TRIGGER AND GCOL
Issue:
------
This problem occurs under the following conditions:
1) Table t2 has a virtual generated column with a certain
   string function.
2) Table t1 has an AFTER INSERT trigger on table t2. It
   contains multiple statements, some unrelated to t2.
3) While inserting into t1, the trigger fails (due to a
   legitimate error).
4) Try to insert into table t2 and this can result in
   incorrect behavior.

Root cause
----------
Step 3 performs open_table for both tables t1 and t2 (since
t2 is part of the trigger). Due to an error, this step is
interrupted and the relevant tables are closed.

    if (thd->locked_tables_mode <= LTM_LOCK_TABLES ||
        table->query_id == thd->query_id)
    {
      DBUG_ASSERT(table->file);
      if (table->db_stat)
        table->file->extra(HA_EXTRA_DETACH_CHILDREN);
      table->cleanup_gc_items();
      ....
    }

Here cleanup_gc_items() is called for table t1 because
query_id is matching. But cleanup_gc_items() is never
called for t2, since the query_id doesn't match and
the table is in LTM_PRELOCKED mode.

While closing the tables it is necessary to specifically
call cleanup_gc_items(). Failure to call this function will
leave the generated column's expression marked as fixed,
while the related table object has been closed.

In Step 4 the generated column's expression is considered
as fixed (even when it isn't) and can result in incorrect
behavior.

Solution:
---------
Stored Functions (SF) and triggers open all the tables
involved at the beginning. If refix_gc_items is allowed for
all the tables at the beginning of an SF and it exits due
to an error during execution, cleanup_gc_items() for all
the tables in the subsequent statements will not be called.

refix_gc_items should be called only when table is directly
used by the statement or substatement (i.e. right before
the relevant substatement is going to be executed).