Bug#14845851 MYSQL CLUSTER - AUTO INCREMENT NOT RESET WHEN TRUNCATE NON-BINLOGGING SQL NODES
When the injector_thread handles a remote truncate it must reset any shared auto_increment
pre-fetch range. If the global table object is dropped and recreated then the new table instance
might otherwise incorrectly see the old auto_increment range. Thus can occur if the mysqld
is not writing the binlog since it then will not handle drop table events for the Ndb table
being dropped and re-created as part of the truncate operation.
Added a new test case that tests this scenario.