Bug#26529942: NON-DEFAULT UMASK AND UMASK_DIR HAS NO EFFECT               IN MYSQL 5.7
Description: Mysql server is not respecting the values of
the environment variables, UMASK and UMASK_DIR.

Analysis: As per the mysql documentation, if UMASK is set,
mysqld uses ($UMASK | 0600) as the mode for file creation.
Similarly if UMASK_DIR is set, mysqld uses ($UMASK_DIR |
0700) as the base mode for directory creation. But the
function "my_init()" is not setting the variables 'my_umask'
and 'my_umask_dir' correctly if the environment variables,
UMASK and UMASK_DIR is set.

Fix: Variables, 'my_umask and 'my_umask_dir' are correctly
set as per the mysql documentation.