Bug#26636238 : ERROR/INPLACE UPGRADE TO READ-ONLY 8.0: PERFORMANCE SCHEMA INITIALIZATION FAILED
Post push fix:
perfschema.init_pfs_from_dd fails on the ASAN build. The reason for
failure is that server restart returns different error code than
expected.

The failure is due to an existing bug in the innodb initialization.
If InnoDB is partially initialized and it encounters an error, InnoDB
does not do proper cleanup required. In the test case, InnoDB is
started with --innodb-read-only option and encounters error while trying
to resize the log files. The default value for innodb_log_file_size
is different in mysqld and mtr.

Fix the test case by supplying the same value of innodb_log_file_size
while restarting the server with which log files were created.