Bug #26390658   RENAMING A PARTITIONED TABLE DOES NOT UPDATE 		MYSQL.INNODB_TABLE_STATS.
PROBLEM

While renaming table that has partitions , table stats don't get updated
with renamed parition names.
Renaming of a partitioned table in 5.7 is done as part of normal
ha_innobase::rename_table and hence after the partitioned tables are
renamed they are not correctly updated to the stats table since table
name fed to the rename stats table function is only the original table
and not the new names of the partition.

FIX

Have made a call to update the table statistics after each rename of
partitioned table happens.

Reviewed by: bin su<bin.x.su@oracle.com>
RB: 17444