Bug#26389402: Outer join optimized away with user-defined functions
When a stored function is used with table column values as arguments
in a WHERE predicate, its not_null_tables property is falsely set to
a non-empty value. If this predicate is applied to an outer join
operation and one of the arguments is from an inner table of the
outer join, the predicate may be used to convert the outer join to
an inner join.

According to SQL standard, only functions that have the "RETURNS NULL
ON NULL INPUT" property should behave like that. (MySQL does not
currently implement this property.) This function changes semantics
of stored functions to not implement RETURNS NULL ON NULL INPUT.

Some internal functions (like internal_table_rows) need the same
possibility as the stored functions. This bug fix will prepare
not_null_tables functionality for these functions, but will not
actually change the implementation for these functions.

A "null_on_null" property is implemented on the Item_func class.
null_on_null is by default true, but is overridden with false in class
Item_func_sp. Any other function that need the same not_null_tables
property may also override null_on_null in its constructor.

When null_on_null is false, not_null_tables_cache is always set
to 0 when updated.

For some Item_func derived classes that effectively did not implement
RETURNS NULL ON NULL INPUT, the overridden function not_null_tables()
that used to return 0 is replaced with a simple assignment to
null_on_null in the constructor, thus eliminating several overridden
functions.